# back-end-prod/back-end-prod/swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/accounts": {
      "get": {
        "operationId": "AccountController_fetchAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountModel"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/accounts/update/profile": {
      "patch": {
        "operationId": "AccountController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountModel"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/accounts/update/password": {
      "put": {
        "operationId": "AccountController_updatePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/accounts/{accountId}": {
      "get": {
        "operationId": "AccountController_getAccountById",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountModel"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/buyers": {
      "get": {
        "operationId": "BuyerController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuyerModel"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Buyers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "operationId": "BuyerController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerModel"
                }
              }
            }
          }
        },
        "tags": [
          "Buyers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "operationId": "BuyerController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerModel"
                }
              }
            }
          }
        },
        "tags": [
          "Buyers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/buyers/byAccountId": {
      "get": {
        "operationId": "BuyerController_findByAccountId",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerModel"
                }
              }
            }
          }
        },
        "tags": [
          "Buyers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/buyers/{id}": {
      "get": {
        "operationId": "BuyerController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerModel"
                }
              }
            }
          }
        },
        "tags": [
          "Buyers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "operationId": "BuyerController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Buyers"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthEmailLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/email/confirm": {
      "post": {
        "operationId": "AuthController_confirmEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthConfirmEmailDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/forgot/password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset/password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/emailVerificationCode/send": {
      "post": {
        "operationId": "EmailVerificationCodeController_sendVerificationCodeEmail",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Email Verification Code"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/emailVerificationCode/verify": {
      "post": {
        "operationId": "EmailVerificationCodeController_verifyCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Email Verification Code"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AccountModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "isValid": {
            "type": "boolean"
          },
          "accountProfilePicture": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "phoneNumber",
          "country",
          "currency",
          "isValid",
          "accountProfilePicture"
        ]
      },
      "UpdateAccountDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "UpdatePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "password",
          "confirmPassword"
        ]
      },
      "BuyerModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "accountId": {
            "type": "string"
          },
          "interestedIn": {
            "type": "string",
            "enum": [
              "Buy a Business",
              "Invest in a business",
              "Buy a Franchise"
            ],
            "example": "Buy a Business"
          },
          "businessStage": {
            "type": "string",
            "enum": [
              "Ideation (MVP/Beta)",
              "Validation (Early revenue)",
              "Growth",
              "Mature (Steady revenue)"
            ],
            "example": "Growth"
          },
          "preferredIndustry": {
            "example": [
              "IT",
              "Finance"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredCountry": {
            "type": "string",
            "example": "India"
          },
          "preferredCities": {
            "example": [
              "Mumbai",
              "Delhi"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "aboutBuyer": {
            "type": "string"
          },
          "aspectsLookedInto": {
            "type": "string"
          },
          "investmentRange": {
            "type": "string"
          },
          "preferredTimeFrame": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "accountId",
          "interestedIn",
          "businessStage",
          "preferredIndustry",
          "preferredCountry",
          "preferredCities",
          "aboutBuyer",
          "aspectsLookedInto",
          "investmentRange",
          "preferredTimeFrame"
        ]
      },
...
      "AuthEmailLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test1@example.com"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthRegisterDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "test1@example.com"
          },
          "phoneNumber": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "phoneNumber",
          "password",
          "confirmPassword"
        ]
      },
      "AuthConfirmEmailDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "AuthForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "AuthResetPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "confirmPassword",
          "token"
        ]
      },
      "VerifyCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ]
      }
    }
  }
}